Extract ServiceOption React component
To render each individual <option> representing a service in the
schedule finder in an encapsulated way, here we extract a ServiceOption
React component. This will later allow us to render different types of
service differently, by using different but interface-compatible
components.
