+++ apps/site/assets/ts/helpers/__tests__/service-test.ts
@@ -1,5 +1,4 @@
 import {
-  serviceDays,
   serviceDate,
   groupServiceByDate,
   groupByType,
@@ -39,45 +38,6 @@ const serviceWithDate: ServiceWithServiceDate = {
   service_date: "2019-06-25",
   name: "weekday"
 };
-
-describe("serviceDays", () => {
-  it("returns an empty string for weekends", () => {
-    const saturday: Service = {
-      ...service,
-      description: "Saturday schedule",
-      type: "saturday",
-      valid_days: [6]
-    };
-    expect(serviceDays(saturday)).toBe("");
-
-    const sunday: Service = {
-      ...service,
-      description: "Sunday schedule",
-      type: "sunday",
-      valid_days: [7]
-    };
-    expect(serviceDays(sunday)).toBe("");
-  });
-
-  it("returns a single day in parentheses", () => {
-    const friday: Service = { ...service, valid_days: [5] };
-    expect(serviceDays(friday)).toBe("Friday");
-  });
-
-  it("returns consecutive days as Monday - Friday", () => {
-    expect(serviceDays(service)).toBe("Monday - Friday");
-    const monToWed: Service = { ...service, valid_days: [1, 2, 3] };
-    expect(serviceDays(monToWed)).toBe("Monday - Wednesday");
-
-    const wedToFri: Service = { ...service, valid_days: [3, 4, 5] };
-    expect(serviceDays(wedToFri)).toBe("Wednesday - Friday");
-  });
-
-  it("lists all non-consecutive days", () => {
-    const monWedFri: Service = { ...service, valid_days: [1, 3, 5] };
-    expect(serviceDays(monWedFri)).toBe("Monday, Wednesday, Friday");
-  });
-});
 
 const currentService = {
   ...serviceWithDate,
